Bump dependencies
Bump @rancher/shell to the latest release, and force some resolutions to
versions without some vulnerabilities marked as critical (though they are
not expected to affect Rancher Desktop in any meaningful way).

The last release of @rancher/shell, 3.0.4, appears to be broken - the menu
for pods don't pop up.  Using the RC version seems to resolve this.

Signed-off-by: Mark Yen <mark.yen@suse.com>